Fixed issue where shared_ptr wasn't being inited correctly
Could not do assignment of shared ptr inside of SimpleSteeringController constructer. This works, but references aren't incremented correctly. This would cause a segfault because the reference count would always be off by n instantiations.
To fix, I took the SetSteering(this) line out of the constructor and moved it into code. Probably not the best solution but to correctly manage the shared_ptr it probably needs to be done this way.

Also added a keyEventReceiver in the example app so we can hit escape and exit it.

Also made obstacles a shared_ptr rather than pass by reference to handle the case where the obstacles array might be deleted (or modified or copied) externally.